requirements.txt
Flask==3.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.37
Werkzeug==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
click==8.2.1
PyPDF2==3.0.1
PyMuPDF==1.25.5
python-docx==1.1.2
easyocr==1.7.2
torch==2.7.0
torchvision==0.22.0
opencv-python==4.11.0.86
numpy==1.26.4
Pillow==11.2.1
python-bidi==0.6.6
pyclipper==1.3.0.post6
shapely==2.1.0
requests==2.32.3
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
urllib3==2.3.0
firebase-admin==6.8.0
PyJWT==2.10.1
google-auth==2.40.2
google-api-core==2.25.0rc1
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
proto-plus==1.26.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1
cachetools==5.5.2
google-crc32c==1.7.1
google-resumable-media==2.7.2
python-dotenv==1.1.0
gunicorn==23.0.0  